requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==0.7.1
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
SQLAlchemy==2.0.13
typing_extensions==4.5.0
#Werkzeug==2.0.2
zipp==3.15.0
google-cloud-bigquery==3.10.0
psycopg2-binary==2.9.6
psycopg2==2.9.6
google-cloud-secret-manager==2.7.1
WTForms==3.0.1
bcrypt==4.0.1
